# GH-121970: Extract ``pydoc_topics`` into a new extension