… Doc/tools/extensions/pyspecific.py
… Lib/pydoc_data/topics.py