@AA-Turner 2025-01-21
Wonderful! I think I may have been testing with Sphinx 8.1 vs 8.2 (unreleased), as I made a change to how the `:kbd:` role is implemented, which would explain the `Ctrl-C` change.