@AA-Turner 2025-01-21
Ok, let's merge this and then ask @Yhg1s what he thinks about backporting to 3.13 and 3.12. I'm assuming the backports will fail due to the change to `Lib/pydoc_data/topics.py`

A